[정순오] 학생 과제목록, 과제 제출 진행 중

+++ lms/src/main/java/com/example/lms/dto/Assignment.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class Assignment {
 	private int assignmentId;
-	private int courseId;
+	private long courseId;
 	private String assignmentName;
 	private String assignmentContent;
 	private String startdate;
